checkpoint for test fa

+++ flashmodels/datasets/alpaca.py
@@ -174,7 +174,7 @@ class DataCollatorForSupervisedDataset(object):
         return dict(
             input_ids=input_ids,
             labels=labels,
-            attention_mask=input_ids.ne(self.tokenizer.pad_token_id),
+            attention_mask=input_ids.ne(self.tokenizer.pad_token_id).int(),
         )
